Rename Protobuf stuff from 'graph-query-service' to 'graph-query' to bring GQS in line with our other service naming conventions.
