+++ src/proto/graplinc/grapl/api/graph_query/v1beta1/graph_query.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package graplinc.grapl.api.graph_query_service.v1beta1;
+package graplinc.grapl.api.graph_query.v1beta1;
 
 import "graplinc/common/v1beta1/types.proto";
 import "graplinc/grapl/common/v1beta1/types.proto";
