+++ src/proto/graplinc/grapl/api/plugin_sdk/analyzers/v1beta1/analyzers.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package graplinc.grapl.api.plugin_sdk.analyzers.v1beta1;
 
 import "graplinc/common/v1beta1/types.proto";
-import "graplinc/grapl/api/graph_query_service/v1beta1/graph_query_service.proto";
+import "graplinc/grapl/api/graph_query/v1beta1/graph_query.proto";
 import "graplinc/grapl/common/v1beta1/types.proto";
 
 // Represents an update to a node's string property
@@ -76,7 +76,7 @@ message AnalyzerName {
 // An ExecutionHit represents the output of an Analyzer, including
 message ExecutionHit {
   // The graph that the Analyzer matched against as well as any context
-  graplinc.grapl.api.graph_query_service.v1beta1.GraphView graph_view = 1;
+  graplinc.grapl.api.graph_query.v1beta1.GraphView graph_view = 1;
   // Any lenses that the graph should be attached to and scored against
   repeated LensRef lens_refs = 2;
   // The name of the Analyzer that triggered
