+++ src/rust/graph-query-proxy/src/server.rs
@@ -1,5 +1,5 @@
 use rust_proto::{
-    graplinc::grapl::api::graph_query_service::v1beta1::{
+    graplinc::grapl::api::graph_query::v1beta1::{
         client::{
             GraphQueryClient,
             GraphQueryClientError,
