+++ src/rust/graph-query/src/main.rs
@@ -10,7 +10,7 @@ use graph_query::{
 };
 use grapl_tracing::setup_tracing;
 use rust_proto::{
-    graplinc::grapl::api::graph_query_service::v1beta1::server::GraphQueryServiceServer,
+    graplinc::grapl::api::graph_query::v1beta1::server::GraphQueryServiceServer,
     protocol::healthcheck::HealthcheckStatus,
 };
 use scylla::CachingSession;
