+++ src/rust/graph-query/src/property_cache.rs
