+++ src/rust/graph-query/src/short_circuit.rs
